refactored SpectralProfilePlotModel (#111) - removed signals from PropertyLabel and PropertyItems - replaced XML by JSON serialization - refactored tooltips - refactored profile candidate plotting - supports antialiasing added PlotStyle().setLineStyle - fixed plot style previews - supports profile data from different vector layers - added SpectralProfileLayerFieldItem - better support of color, label and filter expressions - less layer pointers, uses layer id and QgsProject instance to get layer instances. - added new standard style
